Refactor three optimizer rules into separate files
