+++ arangod/Aql/OptimizerRulesFeature.cpp
@@ -59,6 +59,9 @@
 #include "Aql/Optimizer/Rule/OptimizePaths.h"
 #include "Aql/Optimizer/Rule/GeoIndex.h"
 #include "Aql/Optimizer/Rule/OptimizeCount.h"
+#include "Aql/Optimizer/Rule/AsyncPrefetch.h"
+#include "Aql/Optimizer/Rule/OptimizeProjections.h"
+#include "Aql/Optimizer/Rule/ScatterInCluster.h"
 #include "Aql/Optimizer/Rule/DistributeInCluster.h"
 #include "Aql/Optimizer/Rule/CollectInCluster.h"
 #include "Aql/Optimizer/Rule/DistributeFilterCalcToCluster.h"
